rewrite permission revoke test as integration test
