+++ cli/tests/057_revoke_permissions.out
@@ -1,0 +1,10 @@
+running 7 tests
+OK     runGranted [WILDCARD]
+OK     readGranted [WILDCARD]
+OK     writeGranted [WILDCARD]
+OK     netGranted [WILDCARD]
+OK     envGranted [WILDCARD]
+OK     pluginGranted [WILDCARD]
+OK     hrtimeGranted [WILDCARD]
+
+test result: OK 7 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out [WILDCARD]
